requirements.txt
requests
beautifulsoup4
newspaper4k
pymongo
python-dotenv
lxml
